fix(core): an unknown key refuses the declaration [no-changelog]
A key kendex does not know was a key it skipped, and every way to write
one is a typing accident. `writse:` beside `writes:` is a package that
declares the three files it puts in .git/hooks and a consent screen that
names none of them, with the installer writing them anyway. Nothing tells
that apart from a package which genuinely writes nothing.

Same rule as every field beside it: refused whole rather than read short,
because a partial disclosure reads as a complete one. It costs a package
nothing. The set is seven keys with no extension point, so a declaration
carrying an eighth is one somebody mistyped.

Pinned in the wrong-shape case, which is the same fail-open by another
route. Without the check it fails.
